extensions: add opengl extension to support classic and strict
Based off the work done by Alan Griffiths:
https://github.com/MirServer/egmde-snap/blob/master/snap/snapcraft.yaml

In my prior attempts to make a classic OpenGL snap, I lacked the
"build-attributes": ["no-patchelf"] requirement for libgl1-mesa-dri and
the crashes were hard to diagnose.

Properly adding OpenGL support for snaps is a non-trivial excercise.
This is an attempt to create a single extension which can support
multiple types of snaps (classic & strict) and do the right thing
for OpenGL as the optimal approach to OpenGL changes over time.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>